Extract AssistTest (assist test utils)
Subclass it for now.
We can change this to composition later if needed.
